+++ packages/mantine/src/theme/plasmaCSSVariablesResolver.ts
@@ -6,7 +6,9 @@ export const plasmaCSSVariablesResolver: CSSVariablesResolver = (theme) => {
         dark: {},
         light: {
             '--mantine-color-default-border': theme.colors.gray[2],
+            '--mantine-color-input-border': theme.colors.gray[3],
             '--mantine-color-error': theme.colors.red[5],
+            '--mantine-color-dimmed': theme.colors.gray[5],
         },
     };
     return result;